refactor: Rename constants.py to _constants.py
