+++ src/ikea_api/_constants.py
