+++ src/ikea_api/_endpoints/auth.py
@@ -9,7 +9,7 @@ from pyppeteer import launch
 from pyppeteer.browser import Browser
 from pyppeteer.page import Page
 
-from ikea_api.constants import DEFAULT_HEADERS, Constants, Secrets
+from ikea_api._constants import DEFAULT_HEADERS, Constants, Secrets
 
 
 def get_guest_token() -> str:
