+++ src/ikea_api/_endpoints/item/item_ingka.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from requests import Session
 
+from ikea_api._constants import Constants, Secrets
 from ikea_api._endpoints.item import generic_item_fetcher
-from ikea_api.constants import Constants, Secrets
 from ikea_api.errors import ItemFetchError
