+++ src/ikea_api/_endpoints/item_pip.py
@@ -1,5 +1,5 @@
 from ikea_api._api import API
-from ikea_api.constants import Constants
+from ikea_api._constants import Constants
 from ikea_api.errors import IkeaApiError
