+++ src/ikea_api/_endpoints/order_capture.py
@@ -4,8 +4,8 @@ import re
 from typing import Any
 
 from ikea_api._api import AuthorizedAPI
+from ikea_api._constants import Constants, Secrets
 from ikea_api._endpoints.cart import Cart
-from ikea_api.constants import Constants, Secrets
 from ikea_api.errors import OrderCaptureError
 from ikea_api.types import CustomResponse, OrderCaptureErrorDict
