+++ src/ikea_api/_endpoints/search.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Any
 
 from ikea_api._api import API
-from ikea_api.constants import Constants
+from ikea_api._constants import Constants
 from ikea_api.types import SearchType
